Updated Build
I got an error trying to build the app. I updated the settings.gradle and gradle.properties files. It was successful.
I am also still focusing on displaying the devices to the user only. Once that works, I will move on to connecting to them.
I added the ID for the ListView that will display the devices, created an array that will hold the list of devices that are found on the network. I also added more code to handle devices being found and added to the list. The list theoretically should dynamically change as those two are happening.
I added android.enableJetifier=true to the gradle.properties file. Along with the useAndroidX line, it solved a lot of issues for me.
At this point, I'm still having issues with the Aar metadata while trying to build the app.