Discovered Devices
I've finally gotten it able to discover smart devices on the network. Currently the way it displays them is unacceptable so I'm working on that now and it will be fixed in the next commit.

Most of the unneeded code is gone. Right now its just the bare minimum needed to discover the devices.