IP address Array
It seems ConnectSDK automatically adds the services to the device with a certain IP address so i'm not going to try to do it myself.
I'm simply checking to see if the IP address exists or not and act appropriately.
I also discovered that some services don't even require any pairing at all!
Within the next few commits, I'm expecting to be able to connect to the device.