Robust Discovery
The discovery for the remote is much more robust now. I've decided to go for a model of not letting the user enter the remote UNLESS a device is connected to.
It greatly simplifies the logic and UI elements for me.
The Discovery methods are working very well.
I moved the deviceKey creation to outside the methods so they all have access to it.
Only onDeviceAdded needs the composite key creation means. Since he's the one the rest rely on for their existence.
The AlertDialog will be implemented next. Lets gooooo!!!!