Alert Dialog Implemented
I've implemented the alert dialog now. It works well.
It's not the best looking. I'm calling it from within the onDeviceReady method. Not sure if that's the best place to call it from.
I'll look through the documentation to see if there's a more appropriate method for that.
It displays the same device after being clicked twice at this point. I'm attributing that to calling it within onDeviceReady.
In the next few commits, I'll implement a system for it to stop discovery/display devices to the user after all services are discovered.
I'll also improve how the alert dialog looks.