Charlemagne's Discovery
I'm going to start using famous people in history to signify a major breakthrough or a 'stable' concept. The concept for this is Discovery.
The discovery and display is stable now. Although the display to the user is still not perfect, it's not crashing the app anymore.
The XML was the reason the app was crashing all that time before. I would have never in a thousand years imagined that XML would be a major headache for me. Thanks to Tabe.CrashHandler for coming in clutch though.