Attempting Connections
We've successfully finally implemented the connection method. Now all we need to do is get it do delay just for a few seconds for ConnectSDK to add the services to the device.
We also made the error handling more robust. Removed the selectedDevice array as well since it doesn't serve any purpose now.
Added the deviceKey array back into the code. Apparently it just doesn't want to not add duplicates without it lol.
Added more logs for either a connection success or failure event.