Concept I
If I have a commit where not much was added to it to improve it, I'll simply name it Concept and increase it by a roman numeral.
Added more comments.
Made the deviceKey much more encryptedy.
For some reason the connection is failing. ConnectSDK has ample time to add in all the services so that's not it. At this point, I'm stumped.