Hannibal's Triumph
HOLY HELL WE DID IT!!!!

It connected!!! At least the listener said it did. I'm going to make the UI that we'll use to test it out now.
Man if this works, I'll be on the cloud 9999999999999999. Holy hell alright. Not gonna get too hopeful now.
Of course I will it works!!!!! Life makes sense again!!!