Home button added
I've added the pioneer button for the first command that will be sent to the TV.
I was able to connect to the TV. Now all I need is to send the first command which will happen in the next commit most likely.
I've gotten the delay to work as well. Now it gives ConnectSDK enough time to add all the services, meaning no more failures.
LETS GOOO!