Concept III
I FIGURED IT OUT.
OK. So here's what happened.
I couldn't figure out how to get the ConnectableDevice object into a dictionary that I can then search through for information regarding the device.
I thought I had tried everything and I was even about to modify the ConnectableDevice class file to add in my own property and use it that way.
Then I saw it! The toJSONObject method! It's basically a dictionary. Meaning I can change the device into a json object, save it and access the data that way!
I also found a getConnectedServiceNames() method. Not sure what it does yet but I think it gets the connected device service name :0
Anyway. With this new discovery, I think I might finally be able to pinpoint why the app isn't connecting to the TV. It's either the protocols really don't support it, so it's not even the apps fault, or I'm not implementing it correctly. Time will tell