Manual service addition
I can finally get the uuid for services using the JSONObject Adiutrix. I'll find a better way to implement it than it currently is at in the code.
For now though, it's a big win. This brings us much closer to understanding what's truly happening under the hood. Once we know that, the sky's the limit.
Signed: Man, I'm tired lol