Service availability
I've finally figured it out! I was so close to just grabbing the whole JSON input from the adiutrix and using REGEX to filter through it and find the service. But the simple yet powerful For loop came in clutch.
Now with that complete, I can add in the services to the devices myself. It seems ConnectSDK doesn't do that so great for me. Who knows maybe I'm wrong and just didn't put it in correctly but I'm going to assign it to the device myself.